+++ packages/generator-teams/src/app/templates/gulpfile.js
@@ -2,6 +2,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
+const gulp = require("gulp");
 const package = require("./package.json");
 const argv = require('yargs').argv;
 const config = require('./gulp.config');
@@ -13,7 +14,7 @@ const env = argv["env"];
 if (env === undefined) {
     require('dotenv').config();
 } else {
-    log(`Using custom .env`);
+    log(`Using custom .env: ${env}`);
     require('dotenv').config({ path: path.resolve(process.cwd(), env) });
 }
 process.env.VERSION = package.version;
@@ -21,6 +22,6 @@ process.env.VERSION = package.version;
 const core = require("yoteams-build-core");
 
 // Initialize core build
-core.setup(require('gulp'), config);
+core.setup(gulp, config);
 
-// Add your custom tasks below
+// Add your custom or override tasks below