+++ packages/generator-teams/src/connector/ConnectorGenerator.ts
@@ -121,17 +121,11 @@ export class ConnectorGenerator extends Generator {
 
 
             Yotilities.addAdditionalDeps([
-                ['node-json-db', '0.7.5'],
-                ["@fluentui/react-northstar", "~0.51.0"],
-                ["react", "^16.8.6"],
-                ["react-dom", "^16.8.6"]
+                ['node-json-db', '0.7.5']
             ], this.fs);
 
             Yotilities.addAdditionalDevDeps([
-                ['@types/node-json-db', '0.0.1'],
-                ["@types/react", "16.8.10"],
-                ["file-loader", "6.1.1"],
-                ["typestyle", "2.0.1"]
+                ['@types/node-json-db', '0.0.1']
             ], this.fs);
 
             // update client.ts